fixed ps issue in long form.Scroll always available.Removed abstract-modal. Demo cleanup
