--- addon/index.js
@@ -1,1 +1,0 @@
-export {default as AbstractModal} from './components/abstract-modal'
