+++ addon/styles/addon.scss
@@ -12,16 +12,15 @@ $form-bkgrnd: rgba(0, 0, 0, 0);
 $modal-box-shadow: rgba(64, 64, 65, .5);
 $scroll-box-shadow: rgba(0, 0, 0, .25);
 
-
 .ps-container {
-  max-height: calc(100vh - #{$modal-padding});
-  overflow: auto;
-  padding: 0;
   position: relative;
+  max-height: calc(100vh - #{$modal-padding});
+  padding: 0;
+  overflow: auto;
 
   .ps-scrollbar-y-rail {
-    margin-bottom: $action-bar-height + $info-bar-border-width;
     margin-top: $info-bar-height + $info-bar-border-width;
+    margin-bottom: $action-bar-height + $info-bar-border-width;
   }
 }
 
@@ -38,42 +37,38 @@ $scroll-box-shadow: rgba(0, 0, 0, .25);
 }
 
 .remodal-wrapper {
-  align-items: center;
   display: flex;
+  align-items: center;
   justify-content: center;
 
   .ember-remodal {
     &.window {
-      background-color: $frost-color-white;
-      border: 0;
-      box-shadow: 0 0 20 $modal-box-shadow;
-      display: flex;
-      flex: 1;
-      flex-direction: column;
-      max-height: calc(100vh - #{$modal-padding});
-      max-width: 550px;
-      overflow: hidden;
-      padding: 0;
       position: relative;
       width: 550px;
+      max-width: 550px;
+      max-height: calc(100vh - #{$modal-padding});
+      padding: 0;
+      border: 0;
+      background-color: $frost-color-white;
+      box-shadow: 0 0 20 $modal-box-shadow;
+      overflow: hidden;
 
       .ember-remodal {
         &.yielded {
           &.content {
             display: flex;
-            flex: 1;
             flex-direction: column;
 
             .input-header {
-              align-items: flex-start;
-              border-bottom: $info-bar-border-width solid $frost-color-lgrey-1;
               display: flex;
-              flex-direction: column;
-              height: $info-bar-height + $info-bar-border-width;
-              padding: $info-bar-padding-top-bottom $info-bar-padding-left-right;
               position: absolute;
               top: 0;
+              flex-direction: column;
+              align-items: flex-start;
               width: 100%;
+              height: $info-bar-height + $info-bar-border-width;
+              padding: $info-bar-padding-top-bottom $info-bar-padding-left-right;
+              border-bottom: $info-bar-border-width solid $frost-color-lgrey-1;
               z-index: 999;
 
               &.header-scrolled {
@@ -81,23 +76,23 @@ $scroll-box-shadow: rgba(0, 0, 0, .25);
                 box-shadow: 0 2px 3px $scroll-box-shadow;
 
                 &:after {
-                  background: $frost-color-white;
-                  bottom: 0;
-                  content: '';
-                  height: $info-bar-height + $info-bar-border-width;
-                  left: 0;
-                  opacity: .95;
                   position: absolute;
-                  right: 0;
                   top: 0;
+                  right: 0;
+                  bottom: 0;
+                  left: 0;
+                  height: $info-bar-height + $info-bar-border-width;
+                  background: $frost-color-white;
+                  content: '';
+                  opacity: .95;
                   z-index: -1;
                 }
               }
 
               .primary-title {
+                padding-bottom: 3px;
                 color: $frost-color-night-2;
                 font-size: $frost-font-m;
-                padding-bottom: 3px;
               }
 
               .sub-title {
@@ -134,21 +129,21 @@ $scroll-box-shadow: rgba(0, 0, 0, .25);
             }
 
             .actions {
-              align-items: center;
-              border-top: 1px solid $frost-color-lgrey-1;
-              bottom: 0;
               display: flex;
-              height: $action-bar-height + $action-bar-border-width;
+              position: absolute;
+              bottom: 0;
+              align-items: center;
               justify-content: flex-end;
+              width: 100%;
+              height: $action-bar-height + $action-bar-border-width;
               margin-top: 20px;
               padding-right: 30px;
-              position: absolute;
-              width: 100%;
+              border-top: 1px solid $frost-color-lgrey-1;
               z-index: 999;
 
               > .frost-button {
-                margin-left: 10px;
                 width: 150px;
+                margin-left: 10px;
               }
 
               &.footer-scrolled {
@@ -156,15 +151,15 @@ $scroll-box-shadow: rgba(0, 0, 0, .25);
                 box-shadow: 0 -2px 3px $scroll-box-shadow;
 
                 &:after {
-                  background: $frost-color-white;
-                  bottom: 0;
-                  content: '';
-                  height: $action-bar-height + $action-bar-border-width;
-                  left: 0;
-                  opacity: .95;
                   position: absolute;
-                  right: 0;
                   top: 0;
+                  right: 0;
+                  bottom: 0;
+                  left: 0;
+                  height: $action-bar-height + $action-bar-border-width;
+                  background: $frost-color-white;
+                  content: '';
+                  opacity: .95;
                   z-index: -1;
                 }
               }
